refactor: more refactoring, improved inter-isolate comms.
